+++ b4n/core/managers/views.rs
@@ -617,9 +617,9 @@ impl ViewsManager {
         } else {
             let resource = context.resources.first().cloned().unwrap_or_default();
             let title = plugin.name.clone();
-            let color = self.app_data.borrow().theme.colors.syntax.yaml.string;
+            let colors = self.app_data.borrow().theme.colors.syntax.yaml.clone();
             let keep_output = plugin.keep_output;
-            let command_id = self.worker.borrow_mut().run_plugin(plugin, context, color);
+            let command_id = self.worker.borrow_mut().run_plugin(plugin, context, colors);
             if keep_output {
                 let view = YamlView::new(
                     Rc::clone(&self.app_data),